ai_service/requirements.txt
fastapi==0.128.8
uvicorn==0.40.0
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.52.1
httpx==0.28.1
python-dotenv==1.2.1
anyio==4.12.1
h11==0.16.0
annotated-types==0.7.0
typing-inspection==0.4.2
typing_extensions==4.15.0